requirements.txt
# requirements.txt

# For the interactive web app interface
streamlit==1.37.0
# For data manipulation (reading CSV/Excel files)
pandas==2.2.2
# For interacting with the OpenAI API
openpyxl==3.1.3
# Required by pandas to handle modern .xlsx Excel files
openai==1.35.7
#For loading environment variables from the .env file
python-dotenv==1.0.1